AI_Researcher_Agent/requirements.txt
langchain
langchain-openai
langchain-community
dashscope
python-dotenv
streamlit
httpx
openai